requirements.txt
django==4.2
python-decouple==3.8
django-location-field==2.7.3
six==1.16.0